+++ libs/extensions-api/index.ts
@@ -8,3 +8,4 @@ export * from './create-extension-element-or-fallback.function';
 export * from './create-extension-class.function';
 export * from './umb-lifecycle.interface';
 export * from './entry-point-extension-initializer';
+export * from './types';
