+++ libs/extensions-api/type-guards/is-manifest-class-instance-type.function.ts
@@ -1,4 +1,4 @@
-import type { ManifestClass, ManifestClassWithClassConstructor } from '@umbraco-cms/backoffice/extensions-registry';
+import type { ManifestClass, ManifestClassWithClassConstructor } from '../types';
 
 export function isManifestClassConstructorType(manifest: unknown): manifest is ManifestClassWithClassConstructor {
 	return typeof manifest === 'object' && manifest !== null && (manifest as ManifestClass).class !== undefined;
