+++ libs/extensions-api/type-guards/is-manifest-classable-type.function.ts
@@ -1,7 +1,7 @@
+import type { ManifestBase, ManifestClass } from '../types';
 import { isManifestJSType } from './is-manifest-js-type.function';
 import { isManifestLoaderType } from './is-manifest-loader-type.function';
 import { isManifestClassConstructorType } from './is-manifest-class-instance-type.function';
-import type { ManifestBase, ManifestClass } from '@umbraco-cms/backoffice/extensions-registry';
 
 export function isManifestClassableType(manifest: ManifestBase): manifest is ManifestClass {
 	return (
