+++ libs/extensions-api/type-guards/is-manifest-element-name-type.function.ts
@@ -1,4 +1,4 @@
-import type { ManifestElement, ManifestElementWithElementName } from '@umbraco-cms/backoffice/extensions-registry';
+import type { ManifestElement, ManifestElementWithElementName } from '../types';
 
 export function isManifestElementNameType(manifest: unknown): manifest is ManifestElementWithElementName {
 	return typeof manifest === 'object' && manifest !== null && (manifest as ManifestElement).elementName !== undefined;
