+++ libs/extensions-api/type-guards/is-manifest-elementable-type.function.ts
@@ -1,7 +1,7 @@
+import type { ManifestElement, ManifestBase } from '../types';
 import { isManifestElementNameType } from './is-manifest-element-name-type.function';
 import { isManifestJSType } from './is-manifest-js-type.function';
 import { isManifestLoaderType } from './is-manifest-loader-type.function';
-import type { ManifestElement, ManifestBase } from '@umbraco-cms/backoffice/extensions-registry';
 
 export function isManifestElementableType<ElementType extends HTMLElement = HTMLElement>(
 	manifest: ManifestBase
