+++ libs/extensions-registry/models/dashboard-collection.model.ts
@@ -1,4 +1,4 @@
-import type { ManifestBase } from '.';
+import type { ManifestBase } from '@umbraco-cms/backoffice/extensions-api';
 
 export interface ManifestDashboardCollection extends ManifestBase {
 	type: 'dashboardCollection';
