+++ libs/extensions-registry/models/header-app.model.ts
@@ -1,4 +1,4 @@
-import type { ManifestElement } from '.';
+import type { ManifestElement } from '@umbraco-cms/backoffice/extensions-api';
 
 /**
  * Header apps are displayed in the top right corner of the backoffice
