+++ libs/extensions-registry/models/health-check.model.ts
@@ -1,4 +1,4 @@
-import type { ManifestElement } from '.';
+import type { ManifestElement } from '@umbraco-cms/backoffice/extensions-api';
 
 export interface ManifestHealthCheck extends ManifestElement {
 	type: 'healthCheck';
