+++ libs/extensions-registry/models/menu-item.model.ts
@@ -1,5 +1,5 @@
 import type { UmbMenuItemExtensionElement } from '../interfaces/menu-item-extension-element.interface';
-import type { ManifestElement } from '.';
+import type { ManifestElement } from '@umbraco-cms/backoffice/extensions-api';
 
 export interface ManifestMenuItem extends ManifestElement<UmbMenuItemExtensionElement> {
 	type: 'menuItem';
