+++ libs/extensions-registry/models/modal.model.ts
@@ -1,4 +1,4 @@
-import type { ManifestElement } from '.';
+import type { ManifestElement } from '@umbraco-cms/backoffice/extensions-api';
 
 export interface ManifestModal extends ManifestElement {
 	type: 'modal';
