+++ libs/extensions-registry/models/repository.model.ts
@@ -1,5 +1,4 @@
-import type { ManifestClass } from '.';
-
+import type { ManifestClass } from '@umbraco-cms/backoffice/extensions-api';
 // TODO: Consider adding a ClassType for this manifest. (Currently we cannot know the scope of a repository, therefor we are going with unknown for now.)
 export interface ManifestRepository extends ManifestClass<unknown> {
 	type: 'repository';
