+++ libs/extensions-registry/models/tree-item.model.ts
@@ -1,5 +1,5 @@
 import { UmbTreeItemExtensionElement } from '../interfaces';
-import type { ManifestElement } from '.';
+import type { ManifestElement } from '@umbraco-cms/backoffice/extensions-api';
 
 export interface ManifestTreeItem extends ManifestElement<UmbTreeItemExtensionElement> {
 	type: 'treeItem';
