+++ libs/extensions-registry/models/tree.model.ts
@@ -1,4 +1,4 @@
-import type { ManifestBase } from '.';
+import type { ManifestBase } from '@umbraco-cms/backoffice/extensions-api';
 
 export interface ManifestTree extends ManifestBase {
 	type: 'tree';
