+++ libs/extensions-registry/models/user-profile-app.model.ts
@@ -1,4 +1,4 @@
-import type { ManifestElement } from '.';
+import type { ManifestElement } from '@umbraco-cms/backoffice/extensions-api';
 
 export interface ManifestUserProfileApp extends ManifestElement {
 	type: 'userProfileApp';
