+++ libs/extensions-registry/models/workspace-editor-view.model.ts
@@ -1,5 +1,5 @@
 import type { UmbWorkspaceEditorViewExtensionElement } from '../interfaces/workspace-editor-view-extension-element.interface';
-import type { ManifestWithView } from '.';
+import type { ManifestWithView } from '@umbraco-cms/backoffice/extensions-api';
 
 export interface ManifestWorkspaceEditorView extends ManifestWithView<UmbWorkspaceEditorViewExtensionElement> {
 	type: 'workspaceEditorView';
