+++ libs/extensions-registry/models/workspace-view-collection.model.ts
@@ -1,4 +1,8 @@
-import type { ManifestWithConditions, ManifestWithView, MetaManifestWithView } from '.';
+import type {
+	ManifestWithConditions,
+	ManifestWithView,
+	MetaManifestWithView,
+} from '@umbraco-cms/backoffice/extensions-api';
 
 export interface ManifestWorkspaceViewCollection
 	extends ManifestWithView,
