+++ libs/tree/tree.context.ts
@@ -1,10 +1,9 @@
 import { Observable, map } from 'rxjs';
 import { UmbPagedData, UmbTreeRepository } from '@umbraco-cms/backoffice/repository';
-import type { ManifestTree } from '@umbraco-cms/backoffice/extensions-registry';
+import { ManifestTree, umbExtensionsRegistry } from '@umbraco-cms/backoffice/extensions-registry';
 import { UmbBooleanState, UmbObserverController } from '@umbraco-cms/backoffice/observable-api';
 import { UmbControllerHostElement } from '@umbraco-cms/backoffice/controller';
 import { createExtensionClass } from '@umbraco-cms/backoffice/extensions-api';
-import { umbExtensionsRegistry } from '@umbraco-cms/backoffice/extensions-registry';
 import { ProblemDetailsModel, TreeItemPresentationModel } from '@umbraco-cms/backoffice/backend-api';
 import { UmbContextProviderController } from '@umbraco-cms/backoffice/context-api';
 import { UmbSelectionManagerBase } from '@umbraco-cms/backoffice/utils';
