+++ src/backoffice/core/components/tree/entity-tree-item/entity-tree-item.element.ts
@@ -4,11 +4,8 @@ import { customElement, property } from 'lit/decorators.js';
 import { UmbEntityTreeItemContext } from './entity-tree-item.context';
 import { UmbLitElement } from '@umbraco-cms/internal/lit-element';
 import { EntityTreeItemResponseModel } from '@umbraco-cms/backoffice/backend-api';
-import {
-	ManifestKind,
-	UmbTreeItemExtensionElement,
-	umbExtensionsRegistry,
-} from '@umbraco-cms/backoffice/extensions-registry';
+import { UmbTreeItemExtensionElement, umbExtensionsRegistry } from '@umbraco-cms/backoffice/extensions-registry';
+import { ManifestKind } from '@umbraco-cms/backoffice/extensions-api';
 
 // TODO: Move to separate file:
 const manifest: ManifestKind = {
