+++ src/backoffice/core/index.ts
@@ -9,8 +9,8 @@ import { UmbBackofficeModalContainerElement } from './components/backoffice-fram
 import { UmbNotificationContext, UMB_NOTIFICATION_CONTEXT_TOKEN } from '@umbraco-cms/backoffice/notification';
 import { UmbModalContext, UMB_MODAL_CONTEXT_TOKEN } from '@umbraco-cms/backoffice/modal';
 import { UmbContextProviderController } from '@umbraco-cms/backoffice/context-api';
-import type { UmbEntryPointOnInit } from '@umbraco-cms/backoffice/extensions-api';
-import { ManifestKind, ManifestTypes } from '@umbraco-cms/backoffice/extensions-registry';
+import type { UmbEntryPointOnInit, ManifestKind } from '@umbraco-cms/backoffice/extensions-api';
+import { ManifestTypes } from '@umbraco-cms/backoffice/extensions-registry';
 
 import './notification';
