+++ src/backoffice/packages/repository/package.store.ts
@@ -4,7 +4,7 @@ import { UmbContextToken } from '@umbraco-cms/backoffice/context-api';
 import { UmbControllerHostElement } from '@umbraco-cms/backoffice/controller';
 import { UmbStoreBase } from '@umbraco-cms/backoffice/store';
 import type { PackageMigrationStatusResponseModel } from '@umbraco-cms/backoffice/backend-api';
-import type { ManifestBase } from '@umbraco-cms/backoffice/extensions-registry';
+import type { ManifestBase } from '@umbraco-cms/backoffice/extensions-api';
 import { UmbArrayState } from '@umbraco-cms/backoffice/observable-api';
 
 export const UMB_PACKAGE_STORE_TOKEN = new UmbContextToken<UmbPackageStore>('UmbPackageStore');
