+++ src/backoffice/templating/components/file-system-tree-item/file-system-tree-item.element.ts
@@ -4,7 +4,8 @@ import { customElement, property } from 'lit/decorators.js';
 import { UmbTreeItemElement } from '../../../core/components/tree/tree-item/tree-item.element';
 import { UmbFileSystemTreeItemContext } from './file-system-tree-item.context';
 import { UmbLitElement } from '@umbraco-cms/internal/lit-element';
-import { ManifestKind, umbExtensionsRegistry } from '@umbraco-cms/backoffice/extensions-registry';
+import { umbExtensionsRegistry } from '@umbraco-cms/backoffice/extensions-registry';
+import { ManifestKind } from '@umbraco-cms/backoffice/extensions-api';
 import { FileSystemTreeItemPresentationModel } from '@umbraco-cms/backoffice/backend-api';
 
 // TODO: Move to separate file:
